Do not check that the file got smaller
I know the purpose of this plugin is to optimize images, and that
usually means making them smaller.

But some plugins, in particular SVGO, sometimes make the image larger on
purpose. In particular, some of *its* plugins add data to the SVG. For
example, the `prefixIds` plugin prefixes class names and IDs in the SVG
so that many SVGs with similar IDs on their elements can be used inline
in a single document without conflicting with each other. This
necessarily adds something to the filesize.

Often the amount added is less than the amount taken away by other
optimizations, and so everything seems fine. But sometimes the input SVG
was already sufficiently optimized that more is being added than taken
away.

In these cases, imagemin-webpack-plugin was discarding the
optimizations, and the original image was left alone. This can lead to
all sorts of bugs (IDs colliding, elements within the SVG not being
addressable since you expected them to be prefixed, accessibility
attributes not having been added to the SVG element, wrong display size
due to viewBox not being added, ...), and these are very frustrating and
difficult to trace.

imagemin-webpack-plugin shouldn't care about whether the image got
smaller. That corcern lies solely with the individual optimization tools
like SVGO, Optipng, etc. Instead of testing for a smaller file in
imagemin-webpack-plugin, it's better to file bugs against any of those
tools which sometimes output larger files even when undesirable.